ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
torch==2.1.1
transformers==4.35.2
accelerate==0.24.1
bitsandbytes==0.41.1
scipy==1.11.4
openai-whisper==20231117
soundfile==0.12.1
numpy==1.24.3
python-multipart==0.0.6 